pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "seidr-smidja"
version = "0.1.0.dev0"
description = "An agent-only VRM avatar smithy — a headless programmatic forge for AI agents."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Volmarr Wyrd", email = "volmarrwyrd@gmail.com" },
]
keywords = ["vrm", "avatar", "blender", "mcp", "ai-agent", "vtuber", "vrchat"]

dependencies = [
    # Spec layer (Loom)
    "pydantic>=2.0",
    "pyyaml>=6.0",
    # CLI Bridge (Rúnstafr)
    "click>=8.1",
    # REST Bridge (Straumur)
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    # MCP Bridge (Mjöll)
    "mcp>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "httpx>=0.27",   # for Straumur test client
]

[project.scripts]
seidr = "seidr_smidja.bridges.runstafr.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "requires_blender: marks tests that require a live Blender installation (deselect with -m 'not requires_blender')",
]
addopts = "-m 'not requires_blender'"

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
mypy_path = "src"
